+++ jackplay/test-resources/classes/fortest/dynaloaded/CustomLoadedClass.java
@@ -1,0 +1,8 @@
+package fortest.dynaloaded;
+
+public class CustomLoadedClass {
+
+  public String test1(String name) {
+    return "Dyna." + name;
+  }
+}